Cleaned SettingsFragment
