+++ app/src/main/java/com/numero/material_gallery/MaterialGalleryApplication.kt
@@ -1,8 +1,9 @@
 package com.numero.material_gallery
 
 import android.app.Application
-import com.numero.material_gallery.repository.ConfigRepositoryImpl
+import com.google.android.play.core.appupdate.AppUpdateManagerFactory
 import com.numero.material_gallery.repository.ConfigRepository
+import com.numero.material_gallery.repository.ConfigRepositoryImpl
 import org.koin.android.ext.android.get
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
@@ -22,5 +23,6 @@ class MaterialGalleryApplication : Application() {
 
     private val appModule = module {
         single<ConfigRepository> { ConfigRepositoryImpl(androidContext()) }
+        single { AppUpdateManagerFactory.create(androidContext()) }
     }
 }
