+++ app/src/main/res/values/strings.xml
@@ -4,7 +4,6 @@
     <string name="label_settings">Settings</string>
 
     <string name="update_message">"Newer version is available</string>
-    <string name="update_button">Update</string>
 
     <string name="action_search">Search</string>
     <string name="action_settings">@string/label_settings</string>
