+++ build.gradle
@@ -30,7 +30,7 @@ buildscript {
 
                 lifecycle       : [
                         livedata: "androidx.lifecycle:lifecycle-livedata-ktx:2.3.0-rc01",
-                        runtime : "androidx.lifecycle:lifecycle-runtime:2.3.0-rc01"
+                        runtime : "androidx.lifecycle:lifecycle-runtime-ktx:2.3.0-rc01"
                 ]
         ]
 
@@ -39,7 +39,7 @@ buildscript {
         material = "com.google.android.material:material:$material_version"
 
         play = [
-                core: "com.google.android.play:core:1.8.2",
+                core: "com.google.android.play:core:1.9.0",
                 ktx : "com.google.android.play:core-ktx:1.8.1"
         ]
